% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{DATASET_TITLE}
\format{
A data frame with 32 rows and 11 variables:
\describe{
\item{\code{mpg}}{double COLUMN_DESCRIPTION}
\item{\code{cyl}}{double COLUMN_DESCRIPTION}
\item{\code{disp}}{double COLUMN_DESCRIPTION}
\item{\code{hp}}{double COLUMN_DESCRIPTION}
\item{\code{drat}}{double COLUMN_DESCRIPTION}
\item{\code{wt}}{double COLUMN_DESCRIPTION}
\item{\code{qsec}}{double COLUMN_DESCRIPTION}
\item{\code{vs}}{double COLUMN_DESCRIPTION}
\item{\code{am}}{double COLUMN_DESCRIPTION}
\item{\code{gear}}{double COLUMN_DESCRIPTION}
\item{\code{carb}}{double COLUMN_DESCRIPTION}
}
}
\usage{
mtcars
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
