% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean0}
\alias{mean0}
\title{Mean with NA value}
\usage{
mean0(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Mean
}
\description{
Na omit
}
\details{
dsddwadas
}
\examples{
\dontrun{
if(interactive()){
 mean0(c(1, 2, 3, NA))
 }
}
}
