% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean0.R
\name{mean0}
\alias{mean0}
\title{Mean with NA value}
\usage{
mean0(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Mean
}
\description{
Mean NA omit
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
